Narrow one dimensional sizes to 32 bits, in the ordering code.

+++ experiments/ordering/amd1.cpp
@@ -95,7 +95,7 @@ using Graph = std::vector<std::vector<std::int32_t>>;
 struct Cliques {
     std::vector<std::vector<std::int32_t>> members;
     std::vector<bool> live;
-    std::size_t count = 0;
+    std::uint32_t count = 0;
 
     explicit Cliques(std::size_t n) : members(n), live(n, false) {}
     const std::vector<std::int32_t>& at(std::int32_t c) const { return members[c]; }
@@ -156,8 +156,8 @@ struct Buckets {
 // Print a quotient graph: adjacency, incidence, cliques, in the order the
 // structure holds them.
 void amd1Show(const Graph& A, const Graph& I, const Cliques& C,
-              const std::vector<std::size_t>& degrees,
-              const std::vector<std::size_t>& exact, const std::string& title = "",
+              const std::vector<std::uint32_t>& degrees,
+              const std::vector<std::uint32_t>& exact, const std::string& title = "",
               const std::vector<bool>* eliminated = nullptr) {
     const std::size_t n = A.size();
     int width = static_cast<int>(std::to_string(n > 0 ? n - 1 : 0).size());
@@ -211,8 +211,8 @@ void amd1Show(const Graph& A, const Graph& I, const Cliques& C,
 
 // Print the state arrays: degrees, buckets, min degree, members, eliminated,
 // and the order so far.
-void amd1ShowState(const std::vector<std::size_t>& degrees, const Buckets& buckets,
-                  std::size_t minDegree,
+void amd1ShowState(const std::vector<std::uint32_t>& degrees, const Buckets& buckets,
+                  std::uint32_t minDegree,
                   const std::vector<std::vector<std::int32_t>>& superMembers,
                   const std::vector<bool>& eliminated,
                   const std::vector<std::int32_t>& pivots, const std::string& title = "") {
@@ -332,7 +332,7 @@ std::size_t amd1ExactDegree(const Graph& A, const Graph& I, const Cliques& C,
                             const std::vector<std::vector<std::int32_t>>& superMembers,
                             std::vector<std::int32_t>& mark, std::int32_t& tag,
                             std::int32_t u) {
-    std::size_t degree = 0;
+    std::uint32_t degree = 0;
     for (std::int32_t v : amd1Neighbors(A, I, C, mark, tag, u))
         degree += superMembers[v].size();
     return degree;
@@ -450,7 +450,7 @@ amd1Eliminate(Graph& A, Graph& I, Cliques& C, std::vector<bool>& eliminated,
 // everything the picker asks of it. What it does not give is a minimum, which is
 // why minDegree walks. A sorted container would hand over the minimum directly and
 // charge a log on every file, and files outnumber picks.
-void amd1Refile(Buckets& buckets, std::vector<std::size_t>& degrees,
+void amd1Refile(Buckets& buckets, std::vector<std::uint32_t>& degrees,
                std::int32_t u, std::size_t newDegree) {
     buckets.unfile(degrees[u], u);
     degrees[u] = newDegree;
@@ -473,7 +473,7 @@ std::vector<std::int32_t> amd1MinimumDegree(const Graph& G) {
     // removed: a pivot can carry mass-merged vertices out with it, and from mmd1 up
     // an iteration batches several eliminations before one degree update pass. The three
     // counts coincide only where both of those are absent.
-    std::size_t numEliminations = 0;
+    std::uint32_t numEliminations = 0;
     // Summed over the eliminations, |C[p]| being the new clique AFTER the trim, so
     // in supernodal terms the update rather than the front. It is the raw reach of
     // the eliminations, undeduplicated: where a layer deduplicates, the degree
@@ -483,22 +483,22 @@ std::vector<std::int32_t> amd1MinimumDegree(const Graph& G) {
     // Passes of the outer loop, each one a batch of eliminations followed by one
     // degree update pass. Here the batch is always a single elimination, so this
     // equals numEliminations; from mmd1 up the two come apart.
-    std::size_t numIterations = 0;
+    std::uint32_t numIterations = 0;
     std::vector<std::vector<std::int32_t>> superMembers(n);   // for the expansion
     for (std::int32_t u = 0; u < static_cast<std::int32_t>(n); ++u)
         superMembers[u].push_back(u);
     std::vector<bool> eliminated(n, false);
     std::vector<std::int32_t> pivots;             // the order over supervariables
-    std::size_t numEliminatedVertices = 0;                // a counter, not a scan of eliminated
+    std::uint32_t numEliminatedVertices = 0;                // a counter, not a scan of eliminated
     std::size_t nnzL = 0;
 
     // The cache, and the count of degree computations, which is what this layer
     // exists to reduce. Built once, then touched only where it can be wrong.
     // The cache, as in md5, except that from the first elimination it holds a
     // BOUND rather than a degree. exact[] is carried alongside for the trace only.
-    std::vector<std::size_t> degrees(n);          // a degree counts, so it measures
+    std::vector<std::uint32_t> degrees(n);          // a degree counts, so it measures
     for (std::int32_t u = 0; u < static_cast<std::int32_t>(n); ++u) degrees[u] = A[u].size();
-    std::vector<std::size_t> exact = degrees;
+    std::vector<std::uint32_t> exact = degrees;
     // Only the updates are counted. The total, including the initial pass over all
     // n vertices, is that plus n, so the report derives it. That first pass finds
     // |A[u]| with no clique yet formed, which is the bound formula on an empty
@@ -526,7 +526,7 @@ std::vector<std::int32_t> amd1MinimumDegree(const Graph& G) {
     Buckets buckets(n);
     for (std::int32_t u = 0; u < static_cast<std::int32_t>(n); ++u)
         buckets.file(degrees[u], u);
-    std::size_t minDegree = n > 0 ? *std::min_element(degrees.begin(), degrees.end()) : 0;
+    std::uint32_t minDegree = n > 0 ? *std::min_element(degrees.begin(), degrees.end()) : 0;
     std::size_t numBucketProbes = 0;
 
     // |C[c] - C[pivot]| per clique, indexed by clique id, hoisted out of the loop it is used in.
@@ -536,7 +536,7 @@ std::vector<std::int32_t> amd1MinimumDegree(const Graph& G) {
     // ones it reads, so the iteration clears what it wrote rather than the array being rebuilt. The
     // Python twin has no such line, its outside being a dict over the cliques the iteration touched,
     // which is already the right shape.
-    std::vector<std::size_t> outside(n, 0);
+    std::vector<std::uint32_t> outside(n, 0);
 
     // NOT PRODUCTION: display only. The trace is what makes these files teachable and
     // is the whole reason they exist; nothing downstream reads it.
@@ -569,7 +569,7 @@ std::vector<std::int32_t> amd1MinimumDegree(const Graph& G) {
             amd1Eliminate(A, I, C, eliminated, mark, tag, pivot);
         ++numEliminations;
         numCliqueEntries += C[pivot].size();
-        std::size_t degree = neighbors.size();
+        std::uint32_t degree = neighbors.size();
         pivots.push_back(pivot);
         numEliminatedVertices += 1 + mergedVertices.size();
         for (std::int32_t u : mergedVertices) {   // the pivot now stands for them too
@@ -616,7 +616,7 @@ std::vector<std::int32_t> amd1MinimumDegree(const Graph& G) {
         ++tag;
         const std::int32_t inClique = tag;      // membership of C[pivot], one test
         for (std::int32_t v : pivotClique) mark[v] = inClique;
-        std::size_t degme = 0;
+        std::uint32_t degme = 0;
         for (std::int32_t v : pivotClique) degme += superMembers[v].size();
 
         // |C[c] - C[pivot]| ONCE PER CLIQUE. This is the whole reason the bound is
@@ -639,7 +639,7 @@ std::vector<std::int32_t> amd1MinimumDegree(const Graph& G) {
             numMemberVisits += C[c].size();     // what an exact degree pays PER VERTEX
         }
 
-        const std::size_t numLeft = n - numEliminatedVertices;
+        const std::uint32_t numLeft = n - numEliminatedVertices;
         const std::vector<std::int32_t>& refreshedVertices = pivotClique;
         for (std::int32_t u : refreshedVertices) {
             // bound = |A[u]| + |C[pivot] - {u}| + sum |C[c] - C[pivot]| over the
@@ -683,8 +683,8 @@ std::vector<std::int32_t> amd1MinimumDegree(const Graph& G) {
         // member left there is a live vertex standing for itself alone. The first
         // column then holds ext + w - 1 entries below its diagonal, the next
         // ext + w - 2, down to ext, and each column contributes its own diagonal.
-        std::size_t superSize = superMembers[pivot].size();
-        std::size_t externalDegree = C[pivot].size();
+        std::uint32_t superSize = superMembers[pivot].size();
+        std::uint32_t externalDegree = C[pivot].size();
         nnzL += superSize * externalDegree + superSize * (superSize - 1) / 2 + superSize;
 
         std::ostringstream absorbedCliquesText;
